% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_biodiv_modgis.R
\name{f_biodiv_modgis}
\alias{f_biodiv_modgis}
\title{f_biodiv_modgis}
\usage{
f_biodiv_modgis(data, metric, x = "longitude", y = "latitude",
  date = NULL, num_coord = 40, log_transform = TRUE,
  strata = c("id_site", "localite"), plot_var_imp = TRUE)
}
\arguments{
\item{data}{[data frame] The data}

\item{metric}{[string] Column name of the variable to be modeled}

\item{x}{[string] Column name of the longitude}

\item{y}{[string] Column name of the latitude}

\item{date}{[date] Date}

\item{num_coord}{[numeric] Number of coordinates to project}

\item{log_transform}{[boolean] If TRUE, log-transform response}

\item{strata}{[string] Vector of strata}

\item{plot_var_imp}{[boolean] If TRUE, variables importances are plotted}
}
\value{
a Random Forest object
}
\description{
Run the Random Forest model
}
\examples{

# See vignette()

}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.fr}
}
